+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-**2026.26.8**
+**2026.27.0**
 
 **Biography Translation**